+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Release Notes
 * **NEW:** Support for outputting teletext packets to a web socket
 * **NEW:** Support for outputting teletext packets as a T42 file
 * **NEW:** Support for looped output
+* **IMPROVED:** A teletext service can now be selected using the teletext service ID
 * **IMPROVED:** The code has been significantly refactored for improved readability
 * **FIXED:** Out of bounds crash that occurred while decoding enhancement packets
 * **FIXED:** Incorrect processing of header clock if it contained text